fix: harden rendering, recovery, exports, and readiness
Extract cohesive slide, table, Beamer, and web helpers while preserving
existing callable import paths. Confine publication writes and export
reads, validate checkpoints defensively, and isolate web preprocessing
without changing its RenderingError contract.

Align public readiness supervision with the native verifier budget,
reject incomplete aggregate reports, and correct Reveal scroll-mode
width scaling. Add real-I/O regression coverage and update owner
contracts, generated facts, verification receipts, and the PR template.
