+++ docs/_generated/COUNTS.md
@@ -3,7 +3,7 @@
 > Auto-generated by `scripts/docgen/counts.py` from live repo state. Do not edit
 > manually — run `uv run python scripts/docgen/counts.py --write` to refresh.
 
-**Generated from live repo state on 2026-08-10 (UTC).** Volatile literals are re-derived on every run: tracked `infrastructure/` Python-file count via `git ls-files infrastructure | grep .py` (**796**), project-scope + publishing test collection via `pytest --collect-only` (**608** / **823**), the public exemplar roster, and the importable module list. The per-exemplar test/coverage snapshot table is a measured snapshot with source-commit and source-hash provenance in
+**Generated from live repo state on 2026-09-04 (UTC).** Volatile literals are re-derived on every run: tracked `infrastructure/` Python-file count via `git ls-files infrastructure | grep .py` (**804**), project-scope + publishing test collection via `pytest --collect-only` (**624** / **842**), the public exemplar roster, and the importable module list. The per-exemplar test/coverage snapshot table is a measured snapshot with source-commit and source-hash provenance in
 [`coverage_snapshot.json`](coverage_snapshot.json) (see Test Status).
 
 This file aggregates verifiable facts from discovery scripts, CI configuration, and test execution. Human-written documentation should link here rather than duplicate lists or numbers.
@@ -98,7 +98,7 @@ Tracked Python modules (matches the drift gate):
 git ls-files infrastructure | grep -c '\.py$'
 ```
 
-(Last refreshed count: **796** on 2026-08-10 UTC — point-in-time; re-derive with the command above, the literal drifts as the tree changes.)
+(Last refreshed count: **804** on 2026-09-04 UTC — point-in-time; re-derive with the command above, the literal drifts as the tree changes.)
 
 See `infrastructure/AGENTS.md` for module-specific function signatures and entry points.
 
@@ -115,7 +115,7 @@ uv run pytest tests/infra_tests/project/ --collect-only -q --no-cov
 uv run pytest tests/infra_tests/publishing/ --collect-only -q --no-cov
 ```
 
-Result: **608** project-scope infrastructure tests collected and **823** publishing tests collected. Full behavioral gates still live in CI and in the verification commands listed by the relevant `AGENTS.md` files.
+Result: **624** project-scope infrastructure tests collected and **842** publishing tests collected. Full behavioral gates still live in CI and in the verification commands listed by the relevant `AGENTS.md` files.
 
 **Exemplar `pytest --collect-only` totals** (derived live in each project's declared environment; coverage snapshot last updated 2026-08-16; coverage values use the shared `release` test profile in each project's environment):
